Miscellaneous things (sorry for the bad message)
- Add .env and .env.template file with instructions in README
- Start implementing blizzapi wrapper
- Remove raidTeam from raider entity (swagger only)
- Remove eager loading of raiders from raidTeam entity
- Explicitly load raiders in GET endpoints of RaidTeamsController
- Return Not Found if raidTeam / raider does not exist in DELETE endpoints of controllers

+++ backend/src/raid-teams/raid-teams.service.ts
@@ -33,11 +33,15 @@ export class RaidTeamsService {
     }
 
     findAll(): Promise<RaidTeam[]> {
-        return this.raidTeamsRepository.find();
+        return this.raidTeamsRepository.find({
+            relations: ["raiders"],
+        });
     }
 
     findOne(id: string): Promise<RaidTeam> {
-        return this.raidTeamsRepository.findOne(id);
+        return this.raidTeamsRepository.findOne(id, {
+            relations: ["raiders"],
+        });
     }
 
     async rename(id: string, newName: string): Promise<RaidTeam> {
@@ -63,6 +67,11 @@ export class RaidTeamsService {
     }
 
     async remove(id: string): Promise<void> {
+        var raidTeam: RaidTeam = await this.raidTeamsRepository.findOne(id);
+        if (!raidTeam) {
+            throw new RaidTeamNotFoundException(`No raid team with id ${id} exists.`);
+        }
+
         await this.raidTeamsRepository.delete(id);
     }
 }
